+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # NEXT RELEASE
+* New --output-top10 and --almost-top10 args
 * `truncate_to` config forces `show_full_panel=True`
 * `--output-notes` argument and site
 * Auto doublespace numbered lists and letter lists